move new functionality to ux-editor
