+++ frontend/libs/studio-components/src/hooks/useLocalStorage.ts
@@ -34,10 +34,3 @@ const useWebStorage = <T>(
  */
 export const useLocalStorage = <T>(key: string, initialValue?: T) =>
   useWebStorage<T>(typedLocalStorage, key, initialValue);
-
-export const cleanupStaleLocalStorageKeys = (): void => {
-  // Storage keys that are not supported anymore and to be removed should be added in this array
-  const unsupportedStorageKeys = ['selectedFormLayoutSetName'];
-
-  unsupportedStorageKeys.forEach((key) => typedLocalStorage.removeItem(key));
-};
